+++ app/component/itinerary/LegInfo.js
@@ -7,11 +7,12 @@ import Modal from '@hsl-fi/modal';
 import { legShape, configShape } from '../../util/shapes';
 import { legTimeStr } from '../../util/legUtils';
 import { getRouteMode } from '../../util/modeUtils';
-import RouteNumber from '../RouteNumber';
 import { PREFIX_ROUTES, PREFIX_STOPS } from '../../util/path';
 import { getCapacityForLeg } from '../../util/occupancyUtil';
 import Icon from '../Icon';
 import CapacityModal from '../CapacityModal';
+import OnDemandInfo from './OnDemandInfo';
+import RouteNumberContainer from '../RouteNumberContainer';
 
 /* eslint-disable jsx-a11y/click-events-have-key-events, jsx-a11y/no-static-element-interactions */
 export default function LegInfo(
@@ -25,6 +26,8 @@ export default function LegInfo(
     changeHash,
     tabIndex,
     isCallAgency,
+    mobile,
+    isTransitLeg,
   },
   { config, intl },
 ) {
@@ -47,17 +50,41 @@ export default function LegInfo(
 
   const routeNumber = (
     <span aria-hidden="true">
-      <RouteNumber
+      <RouteNumberContainer
+        route={leg.route}
+        className={`line ${mode}`}
         mode={mode}
         alertSeverityLevel={alertSeverityLevel}
-        color={leg.route ? `#${leg.route.color}` : 'currentColor'}
+        color={leg.route.color ? `#${leg.route.color}` : 'currentColor'}
         text={leg.route && leg.route.shortName}
         realtime={false}
         withBar
         fadeLong
+        isTransitLeg={isTransitLeg}
       />
     </span>
   );
+
+  const [infoOpenState, setInfoOpenState] = useState(false);
+  const openOnDemandInfo = () => {
+    setInfoOpenState(true);
+  };
+  const closeOnDemandInfo = () => {
+    setInfoOpenState(false);
+  };
+  if (infoOpenState) {
+    return (
+      <OnDemandInfo
+        routeNumber={routeNumber}
+        route={leg.route}
+        pickupBookingInfo={leg.pickupBookingInfo}
+        agency={leg.route.agency}
+        onClose={closeOnDemandInfo}
+        mobile={mobile}
+      />
+    );
+  }
+
   return (
     <div
       className={cx('itinerary-transit-leg-route', {
@@ -66,18 +93,17 @@ export default function LegInfo(
       })}
     >
       {isCallAgency ? (
-        <> {routeNumber}</>
+        <button type="button" onClick={openOnDemandInfo}>
+          {routeNumber}
+        </button>
       ) : (
         <Link
           onClick={e => {
             e.stopPropagation();
           }}
-          to={
-            `/${PREFIX_ROUTES}/${leg.route.gtfsId}/${PREFIX_STOPS}/${
-              leg.trip.pattern.code
-            }${shouldLinkToTrip ? `/${leg.trip.gtfsId}` : ''}`
-            // TODO: Create a helper function for generating links
-          }
+          to={`/${PREFIX_ROUTES}/${leg.route.gtfsId}/${PREFIX_STOPS}/${
+            leg.trip.pattern.code
+          }${shouldLinkToTrip ? `/${leg.trip.gtfsId}` : ''}`}
           aria-label={`${intl.formatMessage({
             id: mode,
             defaultMessage: 'Vehicle',
@@ -156,6 +182,8 @@ LegInfo.propTypes = {
   changeHash: PropTypes.func,
   tabIndex: PropTypes.number,
   isCallAgency: PropTypes.bool,
+  isTransitLeg: PropTypes.bool,
+  mobile: PropTypes.bool,
 };
 
 LegInfo.defaultProps = {
@@ -164,6 +192,8 @@ LegInfo.defaultProps = {
   alertSeverityLevel: undefined,
   hasNoShortName: undefined,
   isCallAgency: false,
+  isTransitLeg: false,
+  mobile: undefined,
 };
 
 LegInfo.contextTypes = {
