+++ app/component/itinerary/TransitLeg.js
@@ -579,6 +579,8 @@ class TransitLeg extends React.Component {
             changeHash={this.props.changeHash}
             tabIndex={this.props.tabIndex}
             isCallAgency={mode === 'call'}
+            mobile={this.props.mobile}
+            isTransitLeg
           />
 
           {this.state.showAlternativeLegs &&
@@ -596,6 +598,8 @@ class TransitLeg extends React.Component {
                 )}
                 displayTime
                 isCallAgency={mode === 'call'}
+                mobile={this.props.mobile}
+                isTransitLeg
               />
             ))}
           {this.displayAlternativeLegs() && (
@@ -706,6 +710,7 @@ TransitLeg.propTypes = {
   changeHash: PropTypes.func,
   tabIndex: PropTypes.number,
   usingOwnCarWholeTrip: PropTypes.bool,
+  mobile: PropTypes.bool,
 };
 
 TransitLeg.defaultProps = {
@@ -715,6 +720,7 @@ TransitLeg.defaultProps = {
   tabIndex: undefined,
   children: undefined,
   usingOwnCarWholeTrip: false,
+  mobile: undefined,
 };
 
 TransitLeg.contextTypes = {
